feat: added app settings
