+++ StaffManagementApp.sln
@@ -1,0 +1,33 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.31515.178
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "StaffManagementApp", "StaffManagementApp\StaffManagementApp.csproj", "{6BF8DDCA-7050-43C4-8F90-B1B46238E59A}"
+EndProject
+Project("{00D1A9C2-B5F0-4AF3-8072-F6C62B433612}") = "staffManagementDB", "staffManagementDB\staffManagementDB.sqlproj", "{023FC5EF-0149-4F90-9632-5614A044C679}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{6BF8DDCA-7050-43C4-8F90-B1B46238E59A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{6BF8DDCA-7050-43C4-8F90-B1B46238E59A}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{6BF8DDCA-7050-43C4-8F90-B1B46238E59A}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{6BF8DDCA-7050-43C4-8F90-B1B46238E59A}.Release|Any CPU.Build.0 = Release|Any CPU
+		{023FC5EF-0149-4F90-9632-5614A044C679}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{023FC5EF-0149-4F90-9632-5614A044C679}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{023FC5EF-0149-4F90-9632-5614A044C679}.Debug|Any CPU.Deploy.0 = Debug|Any CPU
+		{023FC5EF-0149-4F90-9632-5614A044C679}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{023FC5EF-0149-4F90-9632-5614A044C679}.Release|Any CPU.Build.0 = Release|Any CPU
+		{023FC5EF-0149-4F90-9632-5614A044C679}.Release|Any CPU.Deploy.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {364AAE27-04FA-4CE8-969B-740FB6FD6D7C}
+	EndGlobalSection
+EndGlobal
