+++ StaffManagementApp/App.config
@@ -1,0 +1,6 @@
+﻿<?xml version="1.0" encoding="utf-8" ?>
+<configuration>
+	<appSettings>
+		<add key="ConnectionStringSQLDb" value="Data Source=DESKTOP-6J5224R;Initial Catalog=StaffDB;User ID=sa;Password=sa123"/>
+	</appSettings>
+</configuration>
