+++ StaffManagementApp/staffs/StaffHelper.cs
@@ -9,38 +9,38 @@ namespace StaffManagementApp.staffs {
             switch (staffType) {
                 case nameof(Teacher):
                     Staff teacher = new Teacher();
-                    teacher.AddStaff( staffType);
+                    teacher.AddStaff(staffType);
                     return teacher;
 
                 case nameof(Administrator):
                     Staff administrator = new Administrator();
-                    administrator.AddStaff( staffType);
+                    administrator.AddStaff(staffType);
                     return administrator;
 
                 case nameof(Support):
                     Staff support = new Support();
-                    support.AddStaff( staffType);
+                    support.AddStaff(staffType);
                     return support;
             }
             return null;
         }
-        
+
         public static void StaffUpdate(string staffType) {
-            
-            Staff staff =GetStaff();
+
+            Staff staff = GetStaff();
             if (staff != null) {
 
                 if (staff.StaffType != staffType) {
                     Console.WriteLine("ID not found");
                     return;
                 }
-            staff = staff.UpdateStaff();
-            Sql.DatabaseUpdateStaff(staff);
-             }
+                staff = staff.UpdateStaff();
+                DatabasManagemantSQL.DatabaseUpdateStaff(staff);
+            }
 
         }
 
-        public static void StaffView( string staffType) {
+        public static void StaffView(string staffType) {
             Staff staff = GetStaff();
             if (staff != null) {
                 if (staff.StaffType != staffType) {
@@ -49,25 +49,25 @@ namespace StaffManagementApp.staffs {
                 }
                 staff.ViewStaff();
             }
-            
-           
+
+
         }
 
         private static Staff GetStaff() {
             Console.Write("Enter staff Id:");
             int StaffId = Convert.ToInt32(Console.ReadLine());
-            Staff staff=Sql.DatabaseGetStaff(StaffId);
+            Staff staff = DatabasManagemantSQL.DatabaseGetStaff(StaffId);
             return staff;
         }
 
-        public static void DeleteStaff( string staffType) {
+        public static void DeleteStaff(string staffType) {
             try {
                 Staff staff = GetStaff();
                 if (staff.StaffType != staffType) {
                     Console.WriteLine("Staff belongs to another type");
                     return;
                 }
-                Sql.DatabaseDeleteStaff(staff.StaffId);
+                DatabasManagemantSQL.DatabaseDeleteStaff(staff.StaffId);
                 Console.WriteLine("Deleted");
             }
             catch (Exception) {
