+++ StaffManagementDb/StaffManagementDb.sqlproj
@@ -3,16 +3,16 @@
   <PropertyGroup>
     <Configuration Condition=" '$(Configuration)' == '' ">Debug</Configuration>
     <Platform Condition=" '$(Platform)' == '' ">AnyCPU</Platform>
-    <Name>staffManagementDB</Name>
+    <Name>StaffManagementDb</Name>
     <SchemaVersion>2.0</SchemaVersion>
     <ProjectVersion>4.1</ProjectVersion>
-    <ProjectGuid>{023fc5ef-0149-4f90-9632-5614a044c679}</ProjectGuid>
+    <ProjectGuid>{26f67e31-7fd2-4681-a8e6-743539358c69}</ProjectGuid>
     <DSP>Microsoft.Data.Tools.Schema.Sql.Sql130DatabaseSchemaProvider</DSP>
     <OutputType>Database</OutputType>
     <RootPath>
     </RootPath>
-    <RootNamespace>staffManagementDB</RootNamespace>
-    <AssemblyName>staffManagementDB</AssemblyName>
+    <RootNamespace>StaffManagementDb</RootNamespace>
+    <AssemblyName>StaffManagementDb</AssemblyName>
     <ModelCollation>1033, CI</ModelCollation>
     <DefaultFileStructure>BySchemaAndSchemaType</DefaultFileStructure>
     <DeployToDatabase>True</DeployToDatabase>
@@ -58,12 +58,13 @@
     <Folder Include="Properties" />
     <Folder Include="dbo\" />
     <Folder Include="dbo\Tables\" />
+    <Folder Include="dbo\Functions\" />
     <Folder Include="dbo\Stored Procedures\" />
     <Folder Include="dbo\User Defined Types\" />
-    <Folder Include="dbo\Functions\" />
   </ItemGroup>
   <ItemGroup>
     <Build Include="dbo\Tables\Staff.sql" />
+    <Build Include="dbo\Functions\funcStaffDob.sql" />
     <Build Include="dbo\Stored Procedures\procUpdateStaff.sql" />
     <Build Include="dbo\Stored Procedures\procInsertStaffs.sql" />
     <Build Include="dbo\Stored Procedures\procGetStaff.sql" />
@@ -71,6 +72,5 @@
     <Build Include="dbo\Stored Procedures\procDeleteStaff.sql" />
     <Build Include="dbo\Stored Procedures\procAddStaff.sql" />
     <Build Include="dbo\User Defined Types\Staffs.sql" />
-    <Build Include="dbo\Functions\funcStaffDob.sql" />
   </ItemGroup>
 </Project>
