+++ StaffManagementDb/dbo/User Defined Types/Staffs.sql
@@ -1,8 +1,6 @@
 ﻿CREATE TYPE [dbo].[Staffs] AS TABLE (
-    [StaffName]       NVARCHAR (20) NULL,
+    [StaffName]       VARCHAR (20) NULL,
     [StaffAge]        INT          NULL,
     [StaffType]       VARCHAR (20) NULL,
     [StaffDepartment] VARCHAR (30) NULL);
-
-
